move mock classes in test
