+++ libs/backend-ddd/tests/opportunity/infrastructure/api/brevo/mock/brevoContact.ts
@@ -1,6 +1,6 @@
 import { Result } from 'true-myth'
-import { ContactId } from '../../../../domain/types'
-import { ContactRepository } from '../../../../domain/spi'
+import { ContactId } from '../../../../../../src/opportunity/domain/types'
+import { ContactRepository } from '../../../../../../src/opportunity/domain/spi'
 
 export const addBrevoContactTest: ContactRepository['createOrUpdate'] = async () => {
   return Result.ok(0 as unknown as ContactId) as Result<ContactId, Error>
